code/2022-04-15-testing-a-flink-stateful-function-service/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.8.1
    # via -r requirements.in
aiosignal==1.2.0
    # via aiohttp
apache-flink-statefun==3.2
    # via -r requirements.in
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via
    #   aiohttp
    #   pytest
brotli==1.0.9
    # via geventhttpclient
certifi==2021.10.8
    # via
    #   geventhttpclient
    #   requests
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   requests
click==8.1.2
    # via flask
configargparse==1.5.3
    # via locust
flask==2.1.1
    # via
    #   flask-basicauth
    #   flask-cors
    #   locust
flask-basicauth==0.2.0
    # via locust
flask-cors==3.0.10
    # via locust
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
gevent==21.12.0
    # via
    #   geventhttpclient
    #   locust
geventhttpclient==1.5.3
    # via locust
greenlet==1.1.2
    # via gevent
idna==3.3
    # via
    #   requests
    #   yarl
iniconfig==1.1.1
    # via pytest
itsdangerous==2.1.2
    # via flask
jinja2==3.0.3
    # via
    #   flask
    #   locust
locust==2.8.6
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
msgpack==1.0.3
    # via locust
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
protobuf==3.20.0
    # via apache-flink-statefun
psutil==5.9.0
    # via locust
py==1.11.0
    # via pytest
pyparsing==3.0.8
    # via packaging
pytest==7.1.1
    # via -r requirements.in
pyzmq==22.3.0
    # via locust
requests==2.27.1
    # via locust
roundrobin==0.0.2
    # via locust
six==1.16.0
    # via
    #   flask-cors
    #   geventhttpclient
tomli==2.0.1
    # via pytest
typing-extensions==4.1.1
    # via locust
urllib3==1.26.9
    # via requests
werkzeug==2.1.1
    # via
    #   flask
    #   locust
yarl==1.7.2
    # via aiohttp
zope-event==4.5.0
    # via gevent
zope-interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
